+++ packages/google-cloud-monitoring-dashboards/owlbot.py
@@ -63,3 +63,32 @@ s.move(templated_files, excludes=[".coveragerc"])  # microgenerator has a good .
 s.replace("noxfile.py", '[\"\']-W[\"\']', '# "-W"')
 
 s.shell.run(["nox", "-s", "blacken"], hide_output=False)
+
+# ----------------------------------------------------------------------------
+# Main Branch migration
+# ----------------------------------------------------------------------------
+
+s.replace(
+  "*.rst",
+  "master",
+  "main"
+)
+
+s.replace(
+  "*.rst",
+  "google-cloud-python/blob/main",
+  "google-cloud-python/blob/master"
+)
+
+s.replace(
+  "CONTRIBUTING.rst",
+  "kubernetes/community/blob/main",
+  "kubernetes/community/blob/master"
+)
+
+s.replace(
+  ".kokoro/*",
+  "master",
+  "main"
+)
+